Fix err when ssh_vault_trusted_user_ca_keys empty
Fix the issue causing the role to fail when an empty
ssh_vault_trusted_user_ca_keys list is provided.

Since an empty ssh_vault_trusted_user_ca_keys means that the combined ca
certificates file isn't created, the role errored when it could not find
this file in cases where the list was empty.

Signed-off-by: Jason Rogena <jason@rogena.me>