Use `ssh_trusted_user_ca_keys_file` instead of `sshd_custom_options`
The devsec.hardening.ssh_hardening role template does not set any
default TrustedUserCAKeys so there is no need to remove and replace the
entry.

Removing the entry and adding it at the end causes the CA to not be
picked up in Ubuntu 22.04. This can be tested by running the previos
version of the role before the commit and running `sshd -T` you'll
notice that the TrustedUserCAKeys is set to none